requirement.txt
fastapi
unicorn
sqlalchemy
psycopg2-binary
jwt
bcrypt
pydantic
websockets
anyio
uvicorn[standard]
python-jose[cryptography]
python-dotenv
passlib
jose
starlette